# files-api-sdk/files_api_sdk/models/file_metadata.py
# coding: utf-8

"""
    FastAPI

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json

from datetime import datetime

from pydantic import BaseModel, Field, StrictInt, StrictStr

class FileMetadata(BaseModel):
    """
    Represents a file in S3.  # noqa: E501
    """
    file_path: StrictStr = Field(...)
    last_modified: datetime = Field(...)
    size_bytes: StrictInt = Field(...)
    __properties = ["file_path", "last_modified", "size_bytes"]

    class Config:
        """Pydantic configuration"""
        allow_population_by_field_name = True
        validate_assignment = True

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.dict(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> FileMetadata:
        """Create an instance of FileMetadata from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self):
        """Returns the dictionary representation of the model using alias"""
        _dict = self.dict(by_alias=True,
                          exclude={
                          },
                          exclude_none=True)
        return _dict

    @classmethod
    def from_dict(cls, obj: dict) -> FileMetadata:
        """Create an instance of FileMetadata from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return FileMetadata.parse_obj(obj)

        _obj = FileMetadata.parse_obj({
            "file_path": obj.get("file_path"),
            "last_modified": obj.get("last_modified"),
            "size_bytes": obj.get("size_bytes")
        })
        return _obj