# files-api-sdk/files_api_sdk/models/get_files_response.py
# coding: utf-8

"""
    FastAPI

    No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)

    The version of the OpenAPI document: 0.1.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json


from typing import List, Optional
from pydantic import BaseModel, Field, StrictStr, conlist
from files_api_sdk.models.file_metadata import FileMetadata

class GetFilesResponse(BaseModel):
    """
    Schema for the result of a get files query.  # noqa: E501
    """
    files: conlist(FileMetadata) = Field(...)
    next_page_token: Optional[StrictStr] = Field(...)
    __properties = ["files", "next_page_token"]

    class Config:
        """Pydantic configuration"""
        allow_population_by_field_name = True
        validate_assignment = True

    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.dict(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> GetFilesResponse:
        """Create an instance of GetFilesResponse from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self):
        """Returns the dictionary representation of the model using alias"""
        _dict = self.dict(by_alias=True,
                          exclude={
                          },
                          exclude_none=True)
        # override the default output from pydantic by calling `to_dict()` of each item in files (list)
        _items = []
        if self.files:
            for _item in self.files:
                if _item:
                    _items.append(_item.to_dict())
            _dict['files'] = _items
        # set to None if next_page_token (nullable) is None
        # and __fields_set__ contains the field
        if self.next_page_token is None and "next_page_token" in self.__fields_set__:
            _dict['next_page_token'] = None

        return _dict

    @classmethod
    def from_dict(cls, obj: dict) -> GetFilesResponse:
        """Create an instance of GetFilesResponse from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return GetFilesResponse.parse_obj(obj)

        _obj = GetFilesResponse.parse_obj({
            "files": [FileMetadata.from_dict(_item) for _item in obj.get("files")] if obj.get("files") is not None else None,
            "next_page_token": obj.get("next_page_token")
        })
        return _obj